fix licenses and split tests
